Fix playerToNodeListing null assertion
The assertion is incorrect, when a hosting user unselects a nation and no user is set to play that nation, then the m_playersToNodeListing has a 'null' node value. A precondition checked caught this; scanning for usages we should have fixed any potential NPEs and can remove the precondition check.