Fix necessary/unnecessary unchecked conversion warning
The Eclipse static analyzer complains when @SuppressWarnings("unchecked")
is both present and not present.  Since we're already doing a cast when
iterating over the enumeration, it doesn't really matter what the
generic type argument of the enumeration is.  So just change it to a
wildcard to get rid of the unchecked conversion warning.