Use positive exit status code for failure
Most shells interpret a process' exit status code as an unsigned 8-bit
integer.  Therefore, a negative status code will be mapped to the range
[128,255].  This can result in an aliasing problem.

When a process is terminated via a signal (e.g. SIGKILL), most shells
will set the process' exit status code to, for example, 128 + N, -N
(where again the negative value is mapped to [128,255]), etc.
Therefore, if a process uses a negative status code, the calling shell
cannot distinguish between an application-initiated termination and one
due to a signal.