Fix RMI regression in IRemoteHostUtils
Even static methods can affect the RMI method ordinal calculation.  In
this case, the addtion of newRemoteNameForNode() caused invocations of
the following methods to fail when called from a legacy lobby:

* stopGameHeadlessHostBot()
* shutDownHeadlessHostBot()

To avoid changing the RMI method order, move newRemoteNameForNode() to a
companion object (really a companion class) on the interface.