Co-locate [I]TestDelegateBridge with tests that use it
These types should not be defined in the engine.data package.  At a
minimum, they should be in engine.delegate, as that is where
IDelegateBridge is defined.  However, as the only tests that use them
are in the triplea.delegate package, that's the best place for them now
since they are deprecated, and their use is discouraged.

Minimize the visibility of these deprecated types to further discourage
their use.