Chip away at TestDelegateBridge (#3935)
* Remove unused ITestDelegateBridge methods

* Inline constant constructor argument

* Co-locate [I]TestDelegateBridge with tests that use it

These types should not be defined in the engine.data package.  At a
minimum, they should be in engine.delegate, as that is where
IDelegateBridge is defined.  However, as the only tests that use them
are in the triplea.delegate package, that's the best place for them now
since they are deprecated, and their use is discouraged.

Minimize the visibility of these deprecated types to further discourage
their use.

* Remove setPlayerId()

The player can always be injected via the constructor.

* Minimize calls to setRemote()

* Remove setRemote()

Replace with an override of getRemotePlayer() that returns the mock
ITripleAPlayer.  Modify all tests to configure this mock directly
instead of creating their own.