Add lobby-client with minimal implementation, a first http server and system test
Implementation for right now is super basic, we log a message we get and we log if we throttle. There is quite a bit to continue doing in these areas, notably to actually persist the message, and when we throttle we should only emit a counter metric. We will also need to implement some throttling rules.

The first http server is spark, it was pretty fast to get up and running. A next server will be dropwizard landing in parallel to the http package.

The integration test runs in under 500ms, fires up the server by invoking main, waits and then calls the server and verifies we get a response. The intent of this test is not to verify correctness (that is already verified by the unit tests), but instead to verify that the server works can serve requests and gives reasonable signs of success and correctly shaped output.