Organize/split up documentation
This update splits up the 'dev' documentation into folders and breaks up the README file to a few categories: conventions, process, setup. The 'other' bucket are things at a top level, the idea is to accumulate more files there until it makes sense to create a grouping to organize them once a category emerges. Of notable absence is something like a 'design' and/or a 'concepts' category to describe how parts and the overall system is designed and likewise a concepts folder to describe domain concepts and provide a glossary of terms and definitions.

This update is mostly a break-up of content without much alteration to it