Introduce http-client + http-data
This update continues development of the error-reporting feature and adds a middle-ware piece that provides an API the game-engine client can use to communicate to the http-server. The http client-server interchange format for communication is JSON formatted String which can then be automatically converted to Java objects pretty easily.

The API is developed to a nearly production ready detail, the feature for error-reporting is expanded to support user input to describe the problem to us, which opens up the feature for ad-hoc error-reporting even without a stack trace (so more than just an automated stack trace upload).

Http-client would ideally provide as much validation and safety checks as we can, the entry points are minimal and that helps then guarantee the integrity of the rest of the system. To this extent, the client has a built in throttle to drop messages that wind up being too large (a large arbitrary number was chosen to start to give us an upper bound), and a simplistic rate throttle so only one error report can be sent every few seconds.

The http-client is tested with '[WireMock](http://wiremock.org/)' which is a stand-alone http server simulator (so this update is independent of http-server). Using that we can test failure scenarios very closely.

To avoid re-inventing the wheel http client, [Feign](https://www.baeldung.com/intro-to-feign) is picked up to create an http client implementation that is bound to an API that we define as an interface with annotations. This reduces the need for unit testing to a pretty reasonable minimum as largely we just define an interface and then use a client builder to create the concrete implementation. The burden is then largely for system/integration testing to verify that things would work provided we use the API presented by http-client sub-project.

Next steps:
 - game engine client code update to use the new http-client and have the error reporting feature behind a feature flag.
 - server side updates to comply with the API now defined by the client.
 - infrastructure updates to stand-up the http server