Move JavaFX client to game-headed project (#4022)
* Break GameRunner dependency on JavaFX client

* Move JavaFX client to game-headed project

* Reorganize game-core dependencies

* Run game-headed integration tests from game-core directory

Some integration tests are dependent on loading external files (e.g.
game_engine.properties).  As a temporary workaround until the issue of
external files is properly resolved, we run the game-headed integration
tests with the working directory set to that of the game-core project so
the external files can be located.