Hide default look-and-feel selection logic in LookAndFeel
This required breaking a circular dependency between the initialization
of the ClientSetting and LookAndFeel classes.  The solution was to move
the code in the LookAndFeel class initializer to a new method
(initialize()) that must be called explicitly.  As GameRunner was
already initializing LookAndFeel early on, this seemed like a
no-brainer.