Write intermediate save game to file instead of memory
Writes a save game to a temporary file first so that, in the event of an
error, the user's existing save game file (if present) will not be
overwritten with a partial save game.

This intermediate save was previously done in memory.  However, for very
large save games, this could require allocating a contiguous range of
memory on the order of tens of MiB.