Make backup copies of lobby system properties for lobby watcher
The original lobby system properties need to be cleared (it's not clear
where this requirement is coming from, but it's been in the code for
some time).  However, the original values need to be backed up to
secondary system properties so they can be read at a later time.  For
example, when the lobby watcher is restarted.

This is effectively a revert of #3676, but cleans up the original code a
bit to encapsulate the knowledge of backing up and clearing the lobby
system properties.