Hide AutoCompletion behind JComboBoxBuilder
Removed some unused functionality of JComboBoxBuilder related to saving
and loading ClientSettings.  In order for JComboBoxBuilder to be
generally useful, it needs to support a generic item type.  It's not
clear how to convert such a generic item type from/to a string, as
required by the ClientSetting subsystem.  Since the ClientSetting
support is not used anywhere, I simply punted and removed it for now as
it may be a YAGNI.