Combine http-data module into http-client
In this update http-data module is combined into the http-client module.

It seems to make sense for the client to also own the data contracts
between client/server. It looks like we can have cases where the server
will have other integrations and perhaps will also need http-client for
those communications. In that case the server will need to have a
dependency on http-client, in which case all projects with a dependency
on http-data will also have a dependency on http-client (so they can be
combined).