Do not propagate SQLExceptions to client
When using Postgres, a specific subclass of SQLException is thrown.
Because the client does not depend on the Postgres driver, this subclass
is not available to it, and a ClassNotFoundException is thrown when the
client attempts to deserialize the results of a remote method call that
threw an exception on the server.