Fix NPE in CommentPanel
The methods of ImageFactory are not thread safe.  All calls (other than
the one in CommentPanel) to getSmallFlag() are currently made from the
EDT.  Presumably, the calls from CommentPanel were subject to some kind
of race condition or other thread safety issue that caused
getSmallFlag() to return null in some cases.  The null return value then
caused the NPE.

The fix employed here is to simply move the getSmallFlag() call in
CommentPanel back to the EDT.  This should avoid whatever race condition
caused the null return value.

An analysis of the ImageFactory#getImage(String, boolean) method showed
that it could return null even if throwIfNotFound is true (e.g.
ImageIO#read() may return null).  Therefore, the code in CommentPanel
was further improved to defend against this possibility.