Remove deprecated IsInfantry methods (#4190)
* Remove deprecated IsInfantry methods

m_isInfantry has been deprecated and replaced by m_isLandTransportable.
However, the isInfantry map property must be retained for compatibility
with older maps.  So we simply redirect isInfantry to
isLandTransportable and make the same modifications to m_isInfantry as
done to m_isLandTransportable.  Then calls to getIsInfantry() can be
replaced with getIsLandTransportable().

* Remove m_isInfantry field