Fix Substance color chooser crash (#4365)
* Fix Substance color chooser crash

* Move management of available Substance L&Fs to new service

* Use Substance API to get available L&Fs

* Make Substance L&F manager service presence optional

This is primarily to keep some integration tests in the game-core
project that rely on the client setting framework (e.g.
ClientContextIntegrationTest) from failing.  Because only the
game-headed project provides an implemenation of the required service,
no service provider is available when running those integration tests.

The alternative would be to move those integration tests to the
game-headed project, which didn't seem like the right thing to do at
this time.  Besides, being more resilient to service absence isn't such
a bad thing.