Move UnifiedMessengerHub#waitForNodesToImplement() to test code
This method is only used by test code and should not be part of the
public API of UnifiedMessengerHub.

Now that the method is in test code, I removed the Deprecated
annotation.  While it would be preferable to react to an event rather
than polling the system, we poll in many other places of test code, such
that deprecating this one (now private) method would be inconsistent.