Allow check-in of IDEA run configurations
Included are two initial run configurations files.
Yet to be included would be a third for bot, that is left out of scope
for now.

Changes:
1. Unignore .idea/runConfigurations folder
2. Check-in launchers for: headed runner and lobby

To unignore the '.idea/runConfigurations' folder we must first ignore the children of the '.idea' folder and not the '.idea' folder itself. From the [git-documentation](https://git-scm.com/docs/gitignore)

> "It is not possible to re-include a file if a parent directory of that
file is excluded."

Next, we add a "re-include" rule to add back anything in the
'.idea/runConfigurations' folder, which is where IDEA places its shared
run configuration files. FWIW: It does not seem there is a way to change this
location (which is perhaps just as well as run configurations have not
changed that often after being checked in)

Discussion:
- While an IDEA plugin can be used to read eclipse launcher files, we do
not have a lot of launchers, it's cheap to create the run
configurations, and the plugin is not necessarily seemless to use and
adds extra overhead compared to checking in run-configurations every now
and then. In other words, I've found it's quite a bit easier to manually re-create run
configurations than it is to wire up the plugin. At that rate we would
just as well check in these files.