Refactor: DownloadUtils split into two classes
This refactor is a preliminary step to making DownloadUtils stateful and
slitting up its API. In this change we introduce two new objects that
will eventually be used directly by clients and replace DownloadUtils.
As a stepping stone to this we first create the objects and move the
implementation of DownloadUtils to these objects, and then have
DownloadUtils simply delegate to these new objects.