Add support for offline mute/unmute by username
Need to guard against negative delays when scheduling timer tasks.  It's
possible for the clocks between client and server to be slightly
out-of-sync, which may generate a negative scheduling delay, which will
result in an exception on the server.  (This is even possible currently
when specifying a zero duration to the Quick Unmute command.)  Solution
is to round all negative delays to zero so the task is immediately
scheduled.