Lobby-db documentation, update prod and dev deployment notes
The lobby-db documentation had the pre-infrastructure way to deploy
prod. It's now updated to mention the important pieces in infrastructure
that will deploy DB changes.

Checking how infrastructure deploys, it
[installs FlyWay](https://github.com/triplea-game/infrastructure/blob/master/roles/application/lobby_db/tasks/flyway.sh)
 and executes all commands it needs. The prod deployment scripts removed
here are the old way to deploy prod and are no longer needed.

Second focus of this update is remove any developer notes to
install a local DB and instead focus on documenting the usage
of a docker container that hosts the DB and runs flyway for us.