Wire github API properties from Environment variables
Previously we had placeholders in the github configuration of
which repository to use and with which API key.

This update injects those values with environment variables:
- A support class EnvironmentVariable is added which will fail-on-boot
  if any expected variables are not set.
- Variables are injected to 'ErrorUploadConfiguration'

Given that we use environmetn variables, the 'prod()' infrastructure
that was provided is not necessary. The code will remain the same
between environments, we only need to change which values are provided
in environment variables. Following this thread lead to configuration
classes being collapsed together and simplified.

A try/catch for error is wrapped around the server start now as any
initialization exception is thrown as an error. The try/catch will log
the exception message so we can see that as the first line of output and
saves us the trouble of digging through the exception stack trace
looking for the root cause message.