Add HomePage Link to README and shorten it.
There is a bit too much information on the README page that turns it
into mostly noise. The updates here are to clarify the headers and
remove content.

The license block was large and simply extracted, the game description
list is removed as we go over that on the homepage. The page beginning
with a link to download did not 100% make sense and made scanning a bit
difficult. Instead we now start with a link to the homepage.

Last, the availaibility report title was updated to make it more obvious
what it is.