ErrorReportWindow: add model and drop 'title'
- Remove user input form field 'title' from error report window

- Add a second auto-filled form field for 'additional informaton' on the
report window. This second text area is pre-filled with any error
message or contents from the console. If neither of those, then the text
area is not rendered at all. This second text field is write-enabled to
allow users to update contents if desired.

- Enhancement: Error report window 'submit' button is enabled only when
a minimum amount of text is present.

- Adds a model object to remove logic burden from window class