Decouple Headless and Headed ServerSetupPanelModel
In this refactor we remove the 'extends SetupPanelModel' from
HeadlessServerSetupPanel. This is a setp towards decoupling headed
vs headless panel models.

Note: HeadlessServerSetupPanelModel did not use 'SetupPanelModel' for
all that much in the end, most notably as a 'holder' for a ISetupPanel
reference, which is easily extracted into the child class to then break
the inheritance relationship.