Simplify bot launching
- Simplify: we have some confused code that checks for required and
non-required args, then after that we have some long-time code that
checks the same non-required args are then required. This first check
was completely removed.

- Simplify: non-required args are removed minus bot password (pending
more investigation if it is used). YAGNI on this removal to simplify to
what we use, which is just the required set.

- If the triplea.name does not start with 'Bot', we will now prepend
it first rather than crashing

- remove lobby-hosted-by, we require that value to match 'triplea.name'
exactly, we may as well just use 'triplea.name' only

- comment is now fixed as 'automated_host' and no longer an arg

- server=true is removed and simply hardcoded, same with default
reconnect time

- updated IDEA and eclipse launchers to have updated params