Docs: create legacy folder, move 'engine_code_overview.md'
engine_code_overview.md is a very old document, one of the originals.
The only reason to keep it around and not update it is for historical
reference. A note was added to the top of the documentation that it is
historical/legacy and the documentation file was moved to a new
'legacy/' folder.