Display log level on log level button.
- Primarily we update the console log level button to display the
current log level. This is to provide feedback to the user of what is
selected without having to click on the log level button.

- Text on log level button updated to be short and not have the down
arrow. The down arrow does not always render, hence "Log Level []"
looks a bit confusing, as if there were a 'null' or no log level is set.

- Remove error checking of log level setting. A user cannot set the
value to an arbitrary text, it should be valid.