Remove system checks
Instead of pro-actively alerting the user to a lack of network
connectivity, instead we allow that message only when the user performs
an action that would require network. EG: show error when user clicks
'play online' rather than at launch.

The temp file check is very unlikely to ever catch a problem, between
these two points there is not much value in the system check.

Beyond those points, the messaging and purpose of the system check error
message is confusing, instead of trying to fix it with the above
problems, will instead simply remove them.