Update lobby connection error message (#4605)
The error message is a bit wordy and incorrect in that it says 'will use
backup values where possible'. If we hit an error on this code flow we
will not use backup values and the connect to lobby just fails.

Second, the error message is overly technical/specific in that we first
download a properties file. We can simply say instead 'failed to connect
to lobby' and not give overly technical details that do not help the
user.