Revert "Constant value - HmacSha512 nonce value (#4657)"
This reverts commit 4abe6b6b946f6523304754286fe895c16d58b95d.

nonceLengthInBytes and saltLengthInBytes have no relationship to each
other.  They are completely orthogonal--it's simply coincidental that
they're equal.  Someone coming along later will probably get confused
and think there's a correlation between these two lengths when, in fact,
salt length is actually correlated to the hash length.

The comment associated with saltLengthInBytes is equally important as
the one for nonceLengthInBytes, especially if a future code reader is
doing some kind of security audit.