Do not copy migration files to docker, run flyway instead
Notable change here is that docker no longer runs migrations and flyway
should now be used to do that.

Docker supports having a set of files to run for SQL schema to be
installed. This does not match what we do in production, it also means
flyway errors out when run out of the box as it sees an existing schema.
Instead of having the docker image run migration scripts, we can use
flyway, a convenience script and documentation notes were added.

Also A TODO note was added for desired eventual next step to install
a snapshot so we can have some test data installed out of the box.