Refactoring: Consolidate Messengers API
In short, this refactoring helps hides the components of messengers
behind a single API 'Messengers'. Messengers becomes more of a facade
class for the moment, which is okay as the messengers were
related and those details were not yet well encapsulated.

Next iterations will continue to morph the API of messengers and internal
implementation to help simplify network coding.