Update Feature: No time duration when banning on remote (#4743)
* Update Feature: No time duration when banning on remote

Instead of offering a ban duration on hosts or on bots, bans will be permanent. Given that we'll likely see a host restart, or would want a bot ban to be permanent, it is just as well. Dropping a ban expiry in this case allows both simpler UX and code.

* Use concurrent collection types, remove external locking