Rename lobby database 'ta_users' -> 'lobby'
This update supports the 1.10 release and changes the 1.9 to 1.10 migration strategy.
With a new DB name, we can keep 1.9 running and reading an untouched 'ta_users' database.
When releasing 1.10, we can copy 'ta_users' to 'lobby', deal with flyway baseline issues,
then migrate the new 'lobby' to latest and run a 1.10 lobby with a 1.9 lobby still running in parallel.