Add a documentation file to describe FlyWay Migration file conventions
Flyway migration files are moving to have the lobby version number baked into
them. Because all changes are getting grouped into a single file, this
makes in-flight changes exceptionally likely to conflict with each other.
To avoid this, this document adds a convention where we date-suffix
new migration files to avoid conflicts, and eventually the date-suffixed files
are squashed together.