Simplify mini-ban and mini-muting, name bans become permament (#4762)
* Simplify mini-ban and mini-muting, name bans become permament

On game hosts, username banning does not make sense since it is a name blacklist which the lobby should already be enforcing. Instead we can reserve just the ban by network identifier to keep things simple. The expectation for a game host that bans someone is that the person will not come back, not just that they have to change names. On the lobby side, username bans then become blacklist and permanent. In all cases we remove the mute-by-username to keep just the mute by network identifier.

Change summary:
- Remove mini-ban by username
- Remove mini-mute by username
- Change username blacklisting to be permanent
- Stop tracking moderator IP and mac in username (blacklist) ban table
- Remove mechanism to remove username bans by using a date in the past. Moderator toolbox 2.0 will address how to remove player name blacklists.

* Text update: 'Blacklist Player Name' -> 'Blacklist Username'

* Replace "Player Name" with "Username" to keep terminology consistent

* Additional "Username" naming consistency updates

* Formatting

* Update comment from player name to username