Remove client-side http throttling
Following YAGNI here, the client side throttling is not strictly needed yet. It is also easily defeated, it's main use is buggy-code/error detection. The value we have gotten from it is low, in favor of allowing for other future refators to be easier, this update drops teh client side throttling code.