Fix chat panel not rendering on game staging screen.
Fixes: https://github.com/triplea-game/triplea/issues/4761

Relatively surgical fix of an impossible `instanceof` check condition that prevented the chat panel from rendering The condition was impossible as the class hierarchies did not line up.

One fix considered was to update the instanceof check to check for presumably the right chat model type 'HeadlessChatModel', but this still would have left in the brittle instanceof check. The fix here rolls back to an explicit 'isHeadless' method that was previously used to do this check.