Drop mutes & Simplify
Dropping mutes:
(1) Incomplete feature, game hosts cannot undo mutes. Permanent-only mutes do not make sense. To repro, host a game, mute a new player that joins and notice there is no way to ever undo the mute.
(2) Anyone that is able to mute can also boot or ban, mute is not strictly necessary ('mute' should not be confused with the 'ignore' feature, which remains intact).

In this update, once mutes are dropped, the children of AbstractServerMessenger becomes much simpler and turns out are no longer different. The two children were there to support DB lookups of mute and the chat channel name was used to support mutes.