Simplify ban player API
This update simplifies the internal concept of banning and is not user facing. Instead of having APIs to be used in sequence to check first for a ban by IP and then another API for ban by mac, we can simplify and increase abstraction if we just call a "isPlayerBanned" method and pass IP and mac in one go.