Expose Feign Http Clients / Drop ServiceResponse Wrapper
Before we did not expose feign created http interfaces directly for a few reasons. One was so that we could signal errors through response objects insteads of exceptions, and secondly so that we could wire in transparent behavior like client-side throttling.

It turns out that throwing an exception is a useful simplification, the problem with response objects is that we had a "double" check to first check for a payload, then a second check that the payload was intact and well formed. With this update any time a response is not going to be well formed, we have the server throw an exception to generate a 500. This comes into play with game-client sending a message to lobby-http serve that could get an error talking to github. In that case the lobby-http server now throws an exception instead of returning an empty github link with error message.

\### Change List
- Remove wrapper code around Feign annotated interfaces, use FeignException to signal any error during an http service call.

- Add custom error decoder to attach server http 500 response body to the thrown exception

- switch Feign library Netflix-Feign to open0feign. The open source version has additional APIs for error handling, notably the error decoder.

- Merge FeignFactory.java into HttpClient.java, overall simplifies.

- Drop SendResult.java, it was part of the service response wrappery-code. The intent there was to tell if a mesasge failed client side or with server. It turns out this is too much detail, over-kill, so we'll simply use exceptions.

- Update pattern of creating http clients, now there is a builder method on the feign annotated interfaces. Typically only a URL is needed, the github client is a bit special as it needs credentialss and repo info.

\### Testing updates
- Add wiremock test to test github http client, that was a testing gap.

- Add a library to help wiremock testing, notably some library code to test the error cases and with a helper method to make testing success case easier.