Instead of checking 'isHeadless', request the rendering componet.
This will allow us to decouple a model that is also its own view. After this update the model component returns an optional 'Component' that is the view. Currently it returns 'this', which avoids casting, and would allow us to change the object returned in the future without having to any code that currently uses 'ChatModel'