Let ErrorController throw errors (#4796)
Previous and recent updates changed error handling so that http server or client would throw exceptions on errors. This update fixes a missed update in the error reporting controller. Previously the idea was to return a string representing an error message or to return a proper JSON response. This was a bit complex, and it turns out we can return an http 500 by throwing an exception.

This update fixes error report controller to throw an exception on error and return a 500 in those cases.